+++ src/test/java/org/logic2j/engine/solver/ExtractingSolutionListener.java
@@ -77,7 +77,7 @@ class ExtractingSolutionListener extends CountingSolutionListener {
   }
 
   public void report() {
-    switch ((int) count()) {
+    switch (count()) {
       case 0:
         logger.info("Solving \"{}\" yields no solution", goal);
         break;
